refactor: generator to synthesizer
